+++ src/main/java/console/RoomEscapeApplication.java
@@ -6,8 +6,11 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Scanner;
+import web.domain.Reservation;
+import web.domain.Theme;
 
 public class RoomEscapeApplication {
+
     private static final String ADD = "add";
     private static final String FIND = "find";
     private static final String DELETE = "delete";
