+++ src/main/java/web/domain/Reservation.java
@@ -1,16 +1,17 @@
-package console;
+package web.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
 public class Reservation {
+
     private Long id;
     private LocalDate date;
     private LocalTime time;
     private String name;
     private Theme theme;
 
-    public Reservation(Long id, LocalDate date, LocalTime time, String name, Theme theme) {
+    public Reservation(Long id, LocalDate date, LocalTime time, String name, web.domain.Theme theme) {
         this.id = id;
         this.date = date;
         this.time = time;
@@ -34,7 +35,7 @@ public class Reservation {
         return name;
     }
 
-    public Theme getTheme() {
+    public web.domain.Theme getTheme() {
         return theme;
     }
 }
