+++ src/main/java/web/domain/Theme.java
@@ -1,6 +1,7 @@
-package console;
+package web.domain;
 
 public class Theme {
+
     private String name;
     private String desc;
     private Integer price;
